+++ apps/astarte_appengine_api/test/astarte_appengine_api_web/controllers/device_status_by_group_controller_test.exs
@@ -94,7 +94,8 @@ defmodule Astarte.AppEngine.APIWeb.DeviceStatusByGroupControllerTest do
     "total_received_bytes" => 4_500_000,
     "total_received_msgs" => 45000,
     "previous_interfaces" => @expected_previous_interfaces,
-    "groups" => [@group_name]
+    "groups" => [@group_name],
+    "deletion_in_progress" => false
   }
 
   setup_all do